00._Assignments/03a._Data_Parsing_Servers/02._python/pyproject.toml
[project]
name = "02-python"
version = "0.1.0"
description = ""
authors = [
    {name = "UlrikLn",email = "ulrik9000@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.3,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "02"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
